+++ packages/bootstrap/macros/src/lib.rs
@@ -1,30 +1,20 @@
-use std::{env, path::PathBuf};
-
 use heck::{ToSnakeCase, ToUpperCamelCase};
 use proc_macro::TokenStream;
 use proc_macro2::Span;
 use proc_macro_error::{abort_call_site, proc_macro_error};
 use quote::quote;
-use silkenweb_base::css;
+use silkenweb_base::css::{self, Source};
 use syn::{parse_macro_input, Ident, LitStr};
 
 #[proc_macro]
 #[proc_macro_error]
 pub fn define_icons(input: TokenStream) -> TokenStream {
     let path: LitStr = parse_macro_input!(input);
+    let path = path.value();
+    let css = Source::path(path).unwrap_or_else(|e| abort_call_site!(e));
+    let classes = css::class_names(&css).filter_map(|class| Class::new(&class));
 
-    let root_dir = cargo_manifest_dir();
-    let path = PathBuf::from(root_dir)
-        .join(path.value())
-        .into_os_string()
-        .into_string()
-        .expect("Expected path to be convertible to string");
-
-    let classes = css::class_names(&path)
-        .unwrap_or_else(|e| abort_call_site!("'{}': {}", path, e.to_string()))
-        .filter_map(|class| Class::new(&class));
-
-    code_gen(&path, classes)
+    code_gen(css.dependency(), classes)
 }
 
 struct Class {
@@ -59,7 +49,7 @@ impl Class {
     }
 }
 
-fn code_gen(path: &str, classes: impl Iterator<Item = Class>) -> TokenStream {
+fn code_gen(dependency: &Option<String>, classes: impl Iterator<Item = Class>) -> TokenStream {
     let classes: Vec<_> = classes.collect();
 
     let enum_variants = classes.iter().map(|class| &class.enum_variant);
@@ -83,8 +73,10 @@ fn code_gen(path: &str, classes: impl Iterator<Item = Class>) -> TokenStream {
         )
     });
 
+    let dependency = dependency.iter();
+
     quote!(
-        const _: &[u8] = ::std::include_bytes!(#path);
+        #(const _: &[u8] = ::std::include_bytes!(#dependency);)*
 
         #[derive(Copy, Clone, Eq, PartialEq, Value)]
         pub enum IconType {
@@ -104,12 +96,4 @@ fn code_gen(path: &str, classes: impl Iterator<Item = Class>) -> TokenStream {
         }
     )
     .into()
-}
-
-fn cargo_manifest_dir() -> String {
-    const CARGO_MANIFEST_DIR: &str = "CARGO_MANIFEST_DIR";
-
-    env::var(CARGO_MANIFEST_DIR).unwrap_or_else(|_| {
-        abort_call_site!("Unable to read {} from environment", CARGO_MANIFEST_DIR)
-    })
 }